+++ src/theme/customTheme.ts
@@ -1,8 +1,17 @@
-import { createTheme } from "@mantine/core";
+import { createTheme, virtualColor } from "@mantine/core";
 
 const theme = createTheme({
+  primaryColor: "primary",
+
   defaultRadius: "md",
   cursorType: "pointer",
+  colors: {
+    primary: virtualColor({
+      name: "primary",
+      dark: "cyan",
+      light: "red",
+    }),
+  },
 });
 
 export default theme;